Moved dropzone to different component
